More corrections for Bug - [# 1876] Spoon: rounding problems when writing Number type values to DB
git-svn-id: svn://source.pentaho.org/svnkettleroot/mirror/Kettle/trunk@562 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src/be/ibridge/kettle/core/Const.java
@@ -472,11 +472,11 @@ public class Const
      * @return The rounded floating point value
      */
     
-    public static final float round(double f, int places) 
+    public static final double round(double f, int places) 
     { 
         java.math.BigDecimal bdtemp = new java.math.BigDecimal(f); 
         bdtemp = bdtemp.setScale(places, java.math.BigDecimal.ROUND_HALF_EVEN); 
-        return bdtemp.floatValue(); 
+        return bdtemp.doubleValue(); 
     } 
     
     /* OLD code: caused a lot of problems with very small and very large numbers.
